pyproject.toml
[tool.poetry]
name = "sports-news-website"
version = "0.1.0"
description = "Comprehensive sports news website using Flask and PostgreSQL"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.5"
flask-login = "^0.6.3"
flask-wtf = "^1.2.1"
wtforms = "^3.1.0"
werkzeug = "^3.0.1"
psycopg2-binary = "^2.9.9"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
python-dotenv = "^1.0.0"
apscheduler = "^3.10.4"
pillow = "^10.1.0"
bleach = "^6.1.0"
email-validator = "^2.1.0"
gunicorn = "^21.2.0"
sqlalchemy = "^2.0.41"
trafilatura = "^2.0.0"
unidecode = "^1.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-flask = "^1.3.0"
black = "^23.11.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"